requirements.txt
aiohttp==3.12.15
beautifulsoup4==4.14.3
dash==4.0.0
 
dash-bootstrap-components==2.0.4
edgartools
fredapi==0.5.2
nest-asyncio==1.6.0
numpy==2.4.3
pandas==3.0.1
plotly==6.6.0
pyarrow==23.0.1
python-dotenv==1.2.2
requests==2.32.5
 
yfinance==1.2.0
